+++ test/TypeRegistryGetType.t.sol
@@ -11,7 +11,7 @@ contract TypeRegistryGetTypeTest is SetupTest {
             string memory description,
             string memory specURI,
             address registrant,
-            uint16 version
+            uint96 version
         ) = registry.getType(bytes4(0x12345678));
 
         assertEq(canonicalSignature, "");
@@ -31,7 +31,7 @@ contract TypeRegistryGetTypeTest is SetupTest {
             string memory description,
             string memory specURI,
             address registrant,
-            uint16 version
+            uint96 version
         ) = registry.getType(selector);
 
         assertEq(canonicalSignature, "noArgs()");
@@ -51,7 +51,7 @@ contract TypeRegistryGetTypeTest is SetupTest {
             string memory description,
             string memory specURI,
             address registrant,
-            uint16 version
+            uint96 version
         ) = registry.getType(selector);
 
         assertEq(canonicalSignature, "metadata(address,(string,(uint256,bytes32)[])[],bytes32[2])");
@@ -74,7 +74,7 @@ contract TypeRegistryGetTypeTest is SetupTest {
             string memory description,
             string memory specURI,
             address registrant,
-            uint16 version
+            uint96 version
         ) = registry.getType(selector);
 
         assertEq(canonicalSignature, "mixed(bytes,string,int128)");
